[Bronze V] Title: 사파리월드, Time: 108 ms, Memory: 12856 KB -BaekjoonHub

+++ 백준/Bronze/2420. 사파리월드/사파리월드.java
@@ -4,8 +4,8 @@ public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 
-		long N = sc.nextInt();
-		long M = sc.nextInt();
+		long N = sc.nextLong();
+		long M = sc.nextLong();
 		sc.close();
 		
 		System.out.println(Math.abs(N - M));
